Fix: wikijs redesign

+++ ct/wikijs.sh
@@ -27,32 +27,34 @@ function update_script() {
         msg_error "No ${APP} Installation Found!"
         exit
     fi
-    msg_info "Stopping ${APP}"
-    systemctl stop wikijs
-    msg_ok "Stopped ${APP}"
+    RELEASE=$(curl -s https://api.github.com/repos/Requarks/wiki/releases/latest | grep "tag_name" | awk '{print substr($2, 3, length($2)-4) }')
+    if [[ "${RELEASE}" != "$(cat /opt/${APP}_version.txt)" ]] || [[ ! -f /opt/${APP}_version.txt ]]; then
+        msg_info "Stopping ${APP}"
+        systemctl stop wikijs
+        msg_ok "Stopped ${APP}"
 
-    msg_info "Backing up Data"
-    mkdir -p ~/data-backup
-    cp -R /opt/wikijs/{db.sqlite,config.yml,/data} ~/data-backup
-    msg_ok "Backed up Data"
+        msg_info "Updating ${APP}"
+        mkdir /opt/wikijs-backup
+        cp -R /opt/wikijs/{config.yml,/data} /opt/wikijs-backup
+        rm -rf /opt/wikijs/*
+        cd /opt/wikijs
+        wget -q "https://github.com/requarks/wiki/releases/download/v${RELEASE}/wiki-js.tar.gz"
+        tar -xzf wiki-js.tar.gz
+        cp -R /opt/wikijs-backup/* /opt/wikijs
+        msg_ok "Updated ${APP}"
 
-    msg_info "Updating ${APP}"
-    rm -rf /opt/wikijs/*
-    cd /opt/wikijs
-    wget -q https://github.com/Requarks/wiki/releases/latest/download/wiki-js.tar.gz
-    tar xzf wiki-js.tar.gz
-    msg_ok "Updated ${APP}"
+        msg_info "Starting ${APP}"
+        systemctl start wikijs
+        msg_ok "Started ${APP}"
 
-    msg_info "Restoring Data"
-    cp -R ~/data-backup/* /opt/wikijs
-    rm -rf ~/data-backup
-    $STD npm rebuild sqlite3
-    msg_ok "Restored Data"
-
-    msg_info "Starting ${APP}"
-    systemctl start wikijs
-    msg_ok "Started ${APP}"
-    msg_ok "Updated Successfully"
+        msg_info "Cleaning Up"
+        rm -rf /opt/wikijs/wiki-js.tar.gz
+        rm -rf /opt/wikijs-backup
+        msg_ok "Cleanup Completed"
+        msg_ok "Updated Successfully"
+    else
+        msg_ok "No update required. ${APP} is already at v${RELEASE}"
+    fi
     exit
 }
